Refactor CSS style application and priority handling
- Refactored CSS style application with explicit priority resolution
- Split view property setters into MT and RT variants
- Static CSS applies with priority checks; animations/actions bypass priority
- CSS transitions are unified with the action system
- Added support for direct child selectors (`>`)
- Optimized selector matching using window-level reverse lookup
  to pre-filter views by class before hierarchy checks
- Updated layout, filter, text, and view code to match new style model

+++ src/ui/action/center.cc
@@ -53,8 +53,8 @@ namespace qk {
 	void ActionCenter::addCSSTransition_rt(View *view, CStyleSheets *css) {
 		auto action = KeyframeAction::MakeSSTransition(view, css, css->time(), true);
 		action->retain(); // retain for center
-		action->play_rt();
-		_CSSTransitions_rt.get(uint64_t(view)).push(action);
+		action->play_rt(); // start action
+		_CSSTransitions_rt.get(uintptr_t(view)).push(action);
 	}
 
 	void ActionCenter::removeCSSTransition_rt(View *view) {
